FundooLoginPage/project/requirements.txt
aioredis==1.3.1
amqp==2.5.2
anyjson==0.3.3
asgiref==3.2.5
async-timeout==3.0.1
attrs==19.3.0
autobahn==20.3.1
Automat==20.2.0
billiard==3.6.3.0
celery==4.4.2
certifi==2019.11.28
cffi==1.14.0
channels==2.4.0
channels-redis==2.4.1
chardet==3.0.4
CodeConvert==3.0.2
constantly==15.1.0
coreapi==2.3.3
coreschema==0.0.4
cryptography==2.8
daphne==2.4.1
defusedxml==0.6.0
Django==2.2
django-admin==2.0.1
django-celery==3.3.1
django-celery-beat==1.6.0
django-celery-results==1.2.0
django-cors-headers==3.2.0
django-elasticsearch-dsl==7.1.1
django-elasticsearch-dsl-drf==0.20.5
django-excel-base==1.0.3
django-excel-response2==3.0.1
django-models-ext==1.1.9
django-nine==0.2.3
django-oauth-toolkit==1.2.0
django-redis==4.11.0
django-redis-cache==2.1.0
django-rest-swagger==2.2.0
django-short-url==1.1.4
django-six==1.0.4
django-storages==1.9.1
django-timezone-field==3.1
djangorestframework==3.11.0
djangorestframework-jwt==1.11.0
elasticsearch==7.6.0
elasticsearch-django==6.5
elasticsearch-dsl==7.1.0
furl==2.1.0
hiredis==1.0.1
httplib2==0.17.0
hyperlink==19.0.0
idna==2.9
importlib-metadata==1.5.0
incremental==17.5.0
itypes==1.1.0
Jinja2==2.11.1
kombu==4.6.8
MarkupSafe==1.1.1
msgpack==0.6.2
oauth2==1.9.0.post1
oauthlib==3.1.0
openapi-codec==1.3.2
orderedmultidict==1.0.1
Pillow==5.1.0

psycopg2-binary==2.8.4
pyasn1==0.4.8
pyasn1-modules==0.2.8
pycparser==2.20
pyee==6.0.0
PyHamcrest==2.0.2
PyJWT==1.7.1
pymitter==0.2.3
pyOpenSSL==19.1.0
python-crontab==2.4.0
python-dateutil==2.8.1
python-oauth2==1.1.1
python3-openid==3.1.0
pytz==2019.3
redis==3.4.1
requests==2.23.0
requests-oauthlib==1.3.0
screen==1.0.1
service-identity==18.1.0
shortuuid==1.0.1
simplejson==3.17.0
six==1.14.0
social-auth-app-django==3.1.0
social-auth-core==3.3.0
sqlparse==0.3.1
TimeConvert==1.5.1
Twisted==20.3.0
txaio==20.1.1
tzlocal==2.0.0
Unidecode==1.1.1
uritemplate==3.0.1
urllib3==1.25.8
validate-email==1.3
vine==1.3.0
xlwt==1.3.0
zipp==3.1.0
zope.interface==5.0.0
